fix(runtime): route explicit PR feedback through local review
